상품 상세 페이지 - 비회원 접근 제한

+++ src/main/webapp/WEB-INF/views/user/prod_detail.jsp
@@ -115,9 +115,9 @@
         }
 
         $(document).ready(function (){
-            // 관심상품으로 등록되어 있는지 확인
-            let has = ${hasWish};
-            console.log(has)
+
+            let has = ${hasWish};       // 관심상품으로 등록되어 있는지 확인
+            let userId = ${userId};
 
             // '관심상품' 버튼 상태 표시
             if(has === 1){
@@ -129,25 +129,35 @@
             // 관심 상품 버튼 클릭하면 이벤트 발생
             $("#like-button").on("click", function (e){
 
-                if(has === 0){ // 관심상품 미등록 상태일때 추가
-                    wishListService.add(
-                        {userId: "${userId}",
-                        prodCode: "${gifticon.prodCode}"},
-                        function (result){
-                            $("#like-button").addClass('selected');
-                            has = 1;
-                            alert("관심상품으로 등록되었습니다.");
-                        });
-                }else{ // 관심상품 등록 상태일때 삭제
-                    wishListService.remove(
-                        {userId: "${userId}",
-                        prodCode: "${gifticon.prodCode}"},
-                        function (result){
-                            $("#like-button").removeClass('selected');
-                            has = 0
-                            alert("관심상품에서 삭제되었습니다.");
-                        });
-                }
+                console.log("userId: " +userId);
+                console.log("hasWish: " +has);
+
+
+                if(userId=== 0){    // 로그인이 안 되어 있을 때
+                    alert("로그인 후에 이용가능한 메뉴입니다.")
+                    $(location).attr('href', "/account/signIn");
+
+                }else{
+                    if(has === 0){ // 관심상품 미등록 상태일때 추가
+                        wishListService.add(
+                            {userId: userId,
+                            prodCode: "${gifticon.prodCode}"},
+                            function (result){
+                                $("#like-button").addClass('selected');
+                                has = 1;
+                                alert("관심상품으로 등록되었습니다.");
+                            });
+                    }else{ // 관심상품 등록 상태일때 삭제
+                        wishListService.remove(
+                            {userId: userId,
+                            prodCode: "${gifticon.prodCode}"},
+                            function (result){
+                                $("#like-button").removeClass('selected');
+                                has = 0
+                                alert("관심상품에서 삭제되었습니다.");
+                            });
+                        }
+                    }
             })
         });
